Changes to creating and viewing surveys
Added new checkbox for boolean is_public which is false by default. When survey creator checks the box then the survey is visible on the homepage.
Connected survey view to backend for users to be able to answer surveys and save answers to database.
Removed authentication from home page so public surveys could be viewed.